requirements.txt
# requirements.txt
streamlit
langchain
langchain-community
pypdf
faiss-cpu
ollama
sqlalchemy
psycopg2-binary
pgvector
bcrypt
python-dotenv
streamlit-javascript